Use pg-listen
Also revert renaming abomination that is `Synch`
