+++ package.json
@@ -28,7 +28,8 @@
     "google-protobuf": "^3.11.2",
     "grpc": "^1.24.2",
     "log4js": "^6.1.1",
-    "pg": "^7.18.1",
+    "pg": "^8.1.0",
+    "pg-listen": "^1.5.1",
     "sha3": "^2.1.1",
     "ts-node": "^8.7.0",
     "typescript": "^3.8.3"
