+++ src/index.ts
@@ -1,6 +1,8 @@
 export { Client } from './lib/client';
 export { Contracts, NewContracts } from './lib/contracts'
-export { Synched, NewSynched, VentListener } from './lib/vent';
+export { VentListener } from './lib/vent';
 export { Deploy } from './deploy';
 export { BytesFromString, BytesToString } from './lib/utils';
 export { CallTx } from '@hyperledger/burrow/proto/payload_pb';
+export {SyncContracts} from "./lib/contracts";
+export {NewSync} from "./lib/contracts";
